Added permamnent income category + added possibility to fund a project in saavings tab

+++ BokuDarjan/app/src/main/java/com/example/bokudarjan/envelope/EnvelopeRepository.kt
@@ -7,7 +7,7 @@ import com.example.bokudarjan.expense.ExpenseDAO
 class EnvelopeRepository(private val envelopeDAO: EnvelopeDAO) {
 
     val readAllData: LiveData<List<Envelope>> = envelopeDAO.readAllData()
-    val sumOfEnvelopes: LiveData<Int> = envelopeDAO.getSumOfEnvelopes()
+    val sumOfEnvelopes: LiveData<Float> = envelopeDAO.getSumOfEnvelopes()
 
     suspend fun addUser(envelope: Envelope) {
         envelopeDAO.addEnvelope(envelope)
