+++ BokuDarjan/app/src/main/java/com/example/bokudarjan/expense/ExpenseRepository.kt
@@ -8,8 +8,8 @@ import com.example.bokudarjan.expense.Expense
 class ExpenseRepository(private val expenseDao: ExpenseDAO) {
 
     val readAllData: LiveData<List<Expense>> = expenseDao.readAllData()
-    val sumOfPositiveExpenses: LiveData<Int> = expenseDao.getSumOfPositiveExpenses()
-    val sumOfNegativeExpense: LiveData<Int> = expenseDao.getSumOfNegativeExpenses()
+    val sumOfPositiveExpenses: LiveData<Float> = expenseDao.getSumOfPositiveExpenses()
+    val sumOfNegativeExpense: LiveData<Float> = expenseDao.getSumOfNegativeExpenses()
 
     suspend fun addUser(expense: Expense) {
         expenseDao.addExpense(expense)
